Restructured the Export task to make the json blob reusable.
Now most of the work of Export is done by a mixin which generates
the data for the json blob, and the Export console task just prints
it.

The api of Export hasn't changed, so this should not affect
anything currently depending on export's behavior (ie the pants
intellij plugin).

The motivation for this change is to enable us to build tasks that
generate code in various formats without redoing the work done by
Export. Ideally, we should be able to replace much of the work
done in IdeGen/IdeaGen with references to the data Export creates.